+++ pandapower/converter/opendss/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2016-2026 by University of Kassel and Fraunhofer Institute for Energy Economics
+# and Energy System Technology (IEE), Kassel. All rights reserved.
+
+from pandapower.converter.opendss.from_dss import *